Build: pass SDK version explicitly to compiler calls
In 457cb3cb027e4b115731328559ae374ba5bc27be the problem became apparant
that the target .NET framework _runtime_ version needs to be
defined/controlled in some way to make it possible to cross-compile
Smuxi from Linux to Windows. This commit achieves this in a cleaner way.

Since the C# compiler has a -sdk: parameter that can take the values
3.5, 4.0, 4.5 and so on, we use this to ensure the build output binds
only to method signatures that will be available at runtime on a
different .NET runtime and operating system.

Differently said: with this change it prevents that a build environment
with Mono 6.8 supporting .NET 4.6 won't make the build output dependent
on .NET 4.6, but restricts the runtime requirement to .NET 4.5