Frontend-GNOME: fix quick join/connect crash for Libera.Chat
When trying to connect or join Libera.Chat using the quick join bar in
the top right corner, Smuxi is crashing with:

	Exception Message:
	No protocol manager found for the protocol: smuxi
	Parameter name: protocol

	Exception StackTrace:
	   at Smuxi.Engine.Session.CreateProtocolManager(String protocol)
	   at Smuxi.Engine.Session.Connect(ServerModel server, FrontendManager frontendManager)
	   at Smuxi.Frontend.Gnome.Frontend.TryOpenChatLink(Uri link)
	   at Smuxi.Frontend.Gnome.MenuWidget.OnJoinWidgetActivated(Object sender, EventArgs e)

The reason is that Libera.Chat has a dot "." in the network name and
Frontend.TryOpenChatLink() assumes if the host part of the URI contains
a dot "." then it must be a server name instead of network name.
This is not true with Libera.Chat, despite the dot it is the network name.

The actual URI passed to TryOpenChatLink() is smuxi://libera.chat and a new
server model was constructed with "smuxi" as protocol ID.
Since this scenario can only lead to failure, another safeguard was
added before calling Session.Connect(server, FrontendManager) that checks
that the protocol is not "smuxi".